[更新]分离并完善模板消息支持
